test(windows): share the Git Bash resolver across test packages
Tests that exec a shell script by its native path need Git Bash specifically
on Windows: the WSL launcher at system32\bash.exe is usually first on PATH and
cannot reach Windows paths (C:/Users/... resolves to "No such file or
directory"; it expects /mnt/c/...). A test resolving plain "bash" there either
fails or, worse, passes for the wrong reason.

- internal/testhelpers: add ResolveBashForScripts, which probes the Git for
  Windows install locations before falling back to PATH, and returns the first
  bash on PATH on other platforms.
- cmd/liza/install_script_test.go: drop its private copy in favour of the
  shared helper.
- internal/embedded/enforce_init_hook_test.go: resolve bash the same way and
  pass the hook path in forward-slash form. The helper is duplicated locally
  rather than imported, because internal/testhelpers imports internal/embedded
  (testhelpers/pipeline.go) and these tests live in package embedded, so the
  import would be a cycle.

This is CI infrastructure rather than a fix: on a machine whose PATH already
resolves bash to Git Bash the behaviour is unchanged. It is a prerequisite for
the windows-latest CI job, where PATH offers WSL bash or none.

Measured on windows/amd64: TestInstallScript* pass; the TestEnforceInitHook_*
package has 6 pre-existing Windows failures both before and after this change
(Windows paths inside the hook JSON payloads), addressed separately.
